+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Allow blank NEW_* settings
 * Baseline API
 * Baseline readme.txt tags
+* Centralize styles/scripts handling
 * Change $2 donation request to $5
 * Check for sections() and settings() as needed
 * Comment out init() by default